Moved functions and state from mock provider to TerminateModal
I assume the code is a part of bigger app, so I'd like to shape it in a way that allows easy development and minimal reliance on the rest of the app. I'm only leaving the `user` object, as that's the most likely one to be provided by app's global state.
Also, while technically possible, I prefer not to store functions in state.