Changed componentDidMount() to redirect if no user defined
The way this app should work is, once the account is deleted and user is invalidated, user shouldn't be able to access this part of website. This surely depends on the rest of the app, but for the brievity of this example, we can just redirect user out of this view.