Moved JSX rendered by FeedbackView methods into main JSX block
Using separate methods to generate jsx can help improve readability sometimes and is especially useful  in cases where the jsx has to be manipulated in some way. However, in this case I don't think there are any benefits to use this approach. Most importantly we're hiding the fact that the rendering of CommentForm relies on condition which can cause confusion when reviewing the file.