Simplified Dialog's state management using basic state machine
This was introduced after I was looking for a better way to navigate between the views than existing setNextView & setPreviousView. Simple state machine solved that issue, as well as simplified overall app state control - with just state prop needed to track the progress and status of core app's functionality. Equally, the error messages for each step of the app can be stored in a single state prop.